diagnostico_pesv/requirements.txt
amqp==5.2.0
asgiref==3.8.1
aspectlib==2.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
Django==5.1
django-appconf==1.0.6
django-celery==3.1.17
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-model-utils==4.5.1
django-seed==0.3.1
django-timestampable==1.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
et-xmlfile==1.1.0
Faker==26.0.0
fields==5.0.0
fonttools==4.53.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.2
marshmallow==3.21.3
matplotlib==3.9.1
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfkit==1.0.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pypandoc==1.13
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
# pywin32==308
redis==5.0.8
reportlab==4.2.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
toposort==1.10
tqdm==4.66.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.3
